Q and A sectional Website for college which is used to share the knowledge
A Place to share knowledge and better understand the world. We started “iLearning
Community” where students are gathered at one place on every Saturday and express
themselves to share the knowledge. Students giving the seminars about latest technologies
and explore new ideas. Knowingly or unknowingly if we miss anything then use this website
to fulfil the ambiguities. This is the amazing platform where the students post queries and
understand the things in different dimensions. This is something like a stack overflow.